Average Score Calculator for 3 Students
While this has many limitations, I'm proud of this mini-project because it demonstrates how far I've come starting 0 in only a week. I used AI to produce comments which critques and explains the code, and also to guide me through the process. In the next coming weeks I plan to expand my knowledge in the world of Python by reading more points on W3Schools and also going through exercises on CodeHS. Hopefully in my next projects I will be less dependent on AI and more confident in my skills.